Set logic rewrite to include some testing and PublicKey parameter handling

+++ DSCResources/RS_rsPullCert/RS_rsPullCert.psm1
@@ -9,6 +9,17 @@
   }
 }
 
+Function Get-TargetResource {
+  param (
+    [parameter(Mandatory = $true)][string]$Name,
+    [string]$PublicKey
+  )
+  return @{
+    'Name' = $Name
+    'PublicKey' = $PublicKey
+  }
+}
+
 Function Test-TargetResource {
   param (
     [parameter(Mandatory = $true)][string]$Name,
@@ -47,22 +58,50 @@ Function Set-TargetResource {
     [string]$PublicKey
   )
   $PullServerAddress = (Get-Content $(Join-Path ([Environment]::GetEnvironmentVariable('defaultPath','Machine')) 'secrets.json') -Raw | ConvertFrom-Json).PullServerAddress
+  $SystemCertID = (Get-ChildItem -Path Cert:\LocalMachine\My\ | Where-Object -FilterScript {$_.Subject -eq $("CN=$PullServerAddress")}).Thumbprint
+  $RootCertID = (Get-ChildItem -Path Cert:\LocalMachine\Root\ | Where-Object -FilterScript {$_.Subject -eq $("CN=$PullServerAddress")}).Thumbprint
   $yesterday = (Get-Date).AddDays(-1) | Get-Date -Format MM/dd/yyyy
   
-  Write-Verbose "Creating Pull server certificate..."
-  Get-ChildItem -Path Cert:\LocalMachine\My\ | Where-Object -FilterScript {$_.Subject -eq $("CN=$PullServerAddress")} | Remove-Item
-  & makecert.exe -b $yesterday -r -pe -n $('CN=', $Settings.PullServerAddress -join ''), -ss my, -sr localmachine, -len 2048
+  #check missing certificates
+  if ($SystemCertID -ne $null) {
+    if ($SystemCertID -ne $RootCertID) {
+      CopyPullCertToRootStore -PullServerAddress $PullServerAddress
+    }
+  }
+  else {
+    Write-Verbose "Creating Pull server certificate..."
+    Get-ChildItem -Path Cert:\LocalMachine\My\ | Where-Object -FilterScript {$_.Subject -eq $("CN=$PullServerAddress")} | Remove-Item
+    & makecert.exe -b $yesterday -r -pe -n $("CN=$PullServerAddress"), -ss my, -sr localmachine, -len 2048
+    CopyPullCertToRootStore -PullServerAddress $PullServerAddress
+  }
+
+  # Export Pulic key to a .cer file for use in local pull DSC configuration
+  if ($PublicKey) {
+    if (Test-Path $PublicKey) {
+      $PublicKeyObj = New-Object System.Security.Cryptography.X509Certificates.X509Certificate2
+      $PublicKeyObj.Import($PublicKey)
+      if ($PublicKeyObj.Thumbprint -ne $SystemCertID) {
+        Write-Verbose "Re-exporting Pull server certificate public key to: $PublicKey"
+        (Get-ChildItem Cert:\LocalMachine\My | Where-Object Subject -eq "CN=$PullServerAddress") | Export-Certificate -Type CERT -FilePath $PublicKey -Force
+      }
+    }
+    else {
+      Write-Verbose "Pull certificate public key file not found - exporting it to $PublicKey"
+      (Get-ChildItem Cert:\LocalMachine\My | Where-Object Subject -eq "CN=$PullServerAddress") | Export-Certificate -Type CERT -FilePath $PublicKey -Force
+    }
+  }
+}
+
+Function CopyPullCertToRootStore {
+  param (
+    [parameter(Mandatory = $true)][string]$PullServerAddress
+  )
+  Write-Verbose "Copying Pull server certificate to root store"
   Get-ChildItem -Path Cert:\LocalMachine\Root\ | Where-Object -FilterScript {$_.Subject -eq $("CN=$PullServerAddress")} | Remove-Item
   $store = Get-Item Cert:\LocalMachine\Root
   $store.Open([System.Security.Cryptography.X509Certificates.OpenFlags]'ReadWrite')
   $store.Add( $(New-Object System.Security.Cryptography.X509Certificates.X509Certificate -ArgumentList @(,(Get-ChildItem Cert:\LocalMachine\My | Where-Object Subject -eq "CN=$PullServerAddress").RawData)) )
   $store.Close()
-  
-  # Export Pulic key to a .cer file for use within local pull DSC encryption, if $PublicKey is provided
-  if ($PublicKey) {
-    Write-Verbose "Exporting Pull server certificate public key to: $PublicKey"
-    (Get-ChildItem Cert:\LocalMachine\My | Where-Object Subject -eq "CN=$PullServerAddress") | Export-Certificate -Type CERT -FilePath $PublicKey -Force
-  }
 }
 
 Export-ModuleMember -Function *-TargetResource
